# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/addData": {
            "post": {
                "description": "Create new pdf from scratch",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeneratePDF1"
                ],
                "summary": "Create new pdf from scratch",
                "parameters": [
                    {
                        "type": "string",
                        "name": "address",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "phoneNumber",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "name": "zipAddress",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "products",
                        "name": "products",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "logo",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/addToTemplate": {
            "post": {
                "description": "Create new pdf with positions",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeneratePDF"
                ],
                "summary": "Create new pdf with positions",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "template",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "object",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "object",
                        "description": "phoneNumber",
                        "name": "phoneNumber",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "object",
                        "description": "zipAddress",
                        "name": "zipAddress",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "object",
                        "description": "address",
                        "name": "address",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "object",
                        "description": "logoData",
                        "name": "logoData",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "products",
                        "name": "products",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "logo",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.LogoData": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "models.address": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "models.name": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "models.phoneNumber": {
            "type": "object",
            "properties": {
                "phoneNumber": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                }
            }
        },
        "models.zipAddress": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                },
                "x": {
                    "type": "integer"
                },
                "y": {
                    "type": "integer"
                },
                "zipAddress": {
                    "type": "integer"
                }
            }
        }
    }
}